feat: #20 refactor generate_log and github_api
