+++ Pipfile
@@ -9,6 +9,8 @@ dotenv = "*"
 opencv-python = "*"
 pyserial = "*"
 pytest = "*"
+numpy = "<2.0"
+pybind11 = ">=2.12"
 
 [dev-packages]
