+++ docs/devlog.md
@@ -1,4 +1,4 @@
-# 📘 Development Log (Auto-updated from recent push)
+# ✅ Development Log (Auto-updated from recent push)
 
 
 ----
