+++ scripts/generate_log.py
@@ -1,53 +1,19 @@
 import os
 import re
-import requests
 from datetime import datetime
-from dotenv import load_dotenv
+
+from scripts.config import REPO
+from scripts.github_api import get_latest_commit_message, get_latest_commit_sha, is_merge_commit, \
+    extract_issue_number_from_commit, fetch_issue_detail
 from utils.path_utils import get_target_directory
 
-load_dotenv()
-
-REPO = "seung-gu/smart-labeler"
-GITHUB_TOKEN = os.getenv("GH_TOKEN")
-HEADERS = {
-    "Authorization": f"Bearer {GITHUB_TOKEN}",
-    "Accept": "application/vnd.github+json"
-}
-
-def get_latest_commit_message():
-    return os.popen("git log -1 --pretty=%B").read().strip()
-
-def get_latest_commit_sha():
-    return os.popen("git rev-parse HEAD").read().strip()
-
-def extract_issue_number_from_commit(message):
-    # For merge commits: extract from branch name
-    merge_match = re.match(r"Merge pull request #\d+ from .+/(\d+)-", message)
-    if merge_match:
-        return merge_match.group(1)
-    # For normal commits: extract from #<number>
-    match = re.search(r"#(\d+)", message)
-    return match.group(1) if match else None
-
-def is_merge_commit(message):
-    return re.match(r"Merge pull request #\d+ from .+/\d+-", message) is not None
-
-def fetch_issue_detail(issue_number):
-    url = f"https://api.github.com/repos/{REPO}/issues/{issue_number}"
-    res = requests.get(url, headers=HEADERS)
-    res.raise_for_status()
-    issue = res.json()
-    return {
-        "number": issue["number"],
-        "title": issue["title"],
-        "body": issue.get("body", "")
-    }
 
 def load_log(path):
     return open(path, "r", encoding="utf-8").read() if os.path.exists(path) else ""
 
 def write_devlog(issue, devlog_path):
     devlog_existing = load_log(devlog_path)
+
     today = datetime.now().strftime("%Y-%m-%d")
     date_header = f"## {today}"
     issue_line = f"### - [#{issue['number']}](https://github.com/{REPO}/issues/{issue['number']}) {issue['title']}"
@@ -93,16 +59,21 @@ def write_devlog(issue, devlog_path):
 
 def write_doclog(issue, doclog_path):
     doclog_existing = load_log(doclog_path)
+
     today = datetime.now().strftime("%Y-%m-%d")
     if f"## {today}" not in doclog_existing:
         doclog_existing += f"\n----\n## {today}\n"
+
     # Remove previous block for this issue
     pattern = rf"### - \[#{issue['number']}\]\(.*?\).*?(?=\n### - \[#\d+\]|$)"
     doclog_existing = re.sub(pattern, "", doclog_existing, flags=re.DOTALL).strip()
     doclog_block = [f"### - [#{issue['number']}](https://github.com/{REPO}/issues/{issue['number']}) {issue['title']}"]
+
     if issue.get("body"):
         doclog_block += issue["body"].strip().splitlines()
+
     doclog_block.append("")
+
     with open(doclog_path, "w", encoding="utf-8") as f:
         f.write(doclog_existing.rstrip() + "\n" + "\n".join(doclog_block) + "\n")
     print(f"📚 Doclog updated: #{issue['number']}")
